+++ src/main/java/ApiRest/TurnosRotativos/repository/EmpleadoRepository.java
@@ -7,10 +7,13 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface EmpleadoRepository extends JpaRepository<Empleado, Integer> {
+public interface EmpleadoRepository extends JpaRepository<Empleado, Long> {
 
-    Optional<Empleado> findByNroDocumento(int nroDocumento);
+    Optional<Empleado> findByNroDocumento(Integer nroDocumento);
+
     Optional<Empleado> findByEmail(String email);
-    boolean existsByNroDocumento(int nroDocumento);
+
+    boolean existsByNroDocumento(Integer nroDocumento);
+
     boolean existsByEmail(String email);
 }
