+++ src/main/java/ApiRest/TurnosRotativos/service/ConceptoLaboralService.java
@@ -5,5 +5,6 @@ import ApiRest.TurnosRotativos.dto.ConceptoLaboralDTO;
 import java.util.List;
 
 public interface ConceptoLaboralService {
-    public List<ConceptoLaboralDTO> getConceptosLaborales(Integer id, String nombre);
+
+    public List<ConceptoLaboralDTO> getConceptosLaborales(Long id, String nombre);
 }
