+++ src/main/java/ApiRest/TurnosRotativos/service/EmpleadoService.java
@@ -1,14 +1,18 @@
 package ApiRest.TurnosRotativos.service;
 
 import ApiRest.TurnosRotativos.dto.EmpleadoDTO;
-import ApiRest.TurnosRotativos.entity.Empleado;
 
 import java.util.List;
 
 public interface EmpleadoService {
-    EmpleadoDTO getEmpleado(int id);
+
+    EmpleadoDTO getEmpleado(Long id);
+
     List<EmpleadoDTO> getEmpleados();
+
     EmpleadoDTO createEmpleado(EmpleadoDTO empleadoDTO);
-    EmpleadoDTO updateEmpleado(int id, EmpleadoDTO empleadoDTO);
-    void deleteEmpleado(Integer id);
+
+    EmpleadoDTO updateEmpleado(Long id, EmpleadoDTO empleadoDTO);
+
+    void deleteEmpleado(Long id);
 }
